requirements.txt
discord.py==2.1.0
emoji==2.0.0
types-emoji==2.0.0